fix: 누락된 total count 추가

+++ src/main/java/app/bottlenote/alcohols/dto/response/AlcoholSearchResponse.java
@@ -9,6 +9,6 @@ import java.util.List;
 @Getter
 @AllArgsConstructor(staticName = "of")
 public class AlcoholSearchResponse {
-	private final int totalCount;
+	private final Long totalCount;
 	private final List<AlcoholsSearchDetail> alcohols;
 }
